requirements.txt
asgiref==3.6.0
Django==4.2
django-crispy-forms==1.12.0
django-filter==23.1
sqlparse==0.4.3
